helm: add install/kubernetes/hubble-cli.yaml
In order to debug Relay to Hubble connectivity issues, it can be useful
to have a Pod running with the Hubble CLI.

Relay is based on a scratch image, so kubectl exec'ing into it is not
possible. While the CLI can be found in the Cilium Pods, the mTLS Relay
certificates are not mounted into the Cilium Pods.

This commit introduce a new hubble-cli.yaml file. When experimenting
Relay mTLS issues, it can be used to quickly run a hubble-cli Pod:

    kubectl create -f /path/to/hubble-cli.yaml

Since the Relay mTLS certificates are mounted into the hubble-cli Pod,
one can easily connect to a Hubble server given it's IP address:

    kubectl exec -it -n kube-system hubble-cli -- \
        hubble observe --server tls://${IP?}:4244 \
            --tls-server-name worker0.default.hubble-grpc.cilium.io \
            --tls-ca-cert-files /var/lib/hubble-relay/tls/hubble-server-ca.crt \
            --tls-client-cert-file /var/lib/hubble-relay/tls/client.crt \
            --tls-client-key-file /var/lib/hubble-relay/tls/client.key

Signed-off-by: Alexandre Perrin <alex@kaworu.ch>